Move AddAttributeCommand DeleteAttributeCommand PersonAttributeCommandParser into correct folders
