+++ src/main/java/seedu/address/logic/commands/DeleteAttributeCommand.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.attributes;
+package seedu.address.logic.commands;
 
 import static java.util.Objects.requireNonNull;
