+++ src/test/java/seedu/address/logic/attribute/DeleteAttributeCommandTest.java
@@ -6,7 +6,7 @@ import static seedu.address.testutil.TypicalPersons.getTypicalAddressBook;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.attributes.DeleteAttributeCommand;
+import seedu.address.logic.commands.DeleteAttributeCommand;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
