+++ src/test/java/seedu/address/logic/parser/AddressBookParserTest.java
@@ -14,8 +14,8 @@ import java.util.stream.Collectors;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.attributes.AddAttributeCommand;
-import seedu.address.logic.attributes.DeleteAttributeCommand;
+import seedu.address.logic.commands.AddAttributeCommand;
+import seedu.address.logic.commands.DeleteAttributeCommand;
 import seedu.address.logic.commands.AddCommand;
 import seedu.address.logic.commands.ClearCommand;
 import seedu.address.logic.commands.Command;
